requirements.txt
numpy==2.3.4
opencv_python==4.11.0.86
Pillow==12.0.0
scikit_learn==1.4.2
torch==2.5.1
torchvision==0.20.1
